Fix Chain destructor for types that need dropping
The destructor for Chain would just repeatedly try dropping the same
block of memory over and over. While this doesn't cause any errors, it
does mean there is an infinite loop while trying to dispose of a chunk.

I made the assumption that the pointer is supposed to be advanced to the
next value in the chunk.